calculation logic done but not working

+++ Models/Greeblo.cs
@@ -2,6 +2,6 @@ namespace backend.Models;
 
 public class Greeblo {
     public int Health { get; set; }
-    public int Cacao { get; set; }
+    public double Cacao { get; set; }
     public ChocolateBar[]? Schedule { get; set; }
 }
